requirements.txt
# --- Core Framework & Server ---
fastapi==0.115.8
uvicorn==0.34.0

# --- Async / HTTP ---
aiofiles==24.1.0
aiohttp==3.11.12
httpx==0.28.1
requests==2.32.3

# --- Data Handling ---
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.5
SQLAlchemy==2.0.38
python-dotenv==1.0.1

# --- LangChain & AI Integrations ---
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
langchain-openai==0.3.5
langchain-ollama==0.2.3
langgraph==0.2.72
langgraph-checkpoint==2.0.13
langgraph-sdk==0.1.51
langsmith==0.3.8
openai==1.62.0
tiktoken==0.8.0
orjson==3.10.15
pydantic==2.10.6
pydantic-settings==2.7.1

# --- Document Processing & Generation ---
beautifulsoup4==4.13.3
lxml==5.3.1
unstructured==0.16.20
unstructured-client==0.29.0
markdown2==2.5.3
python-docx==1.1.2
pypdf==5.3.0
PyMuPDF==1.25.3
weasyprint==64.0
htmldocx==0.0.6

# --- Scraping & Parsing ---
Scrapy==2.12.0
fake-useragent==2.0.3
tldextract==5.1.3

# --- Text / NLP ---
nltk==3.9.1
regex==2024.11.6
langdetect==1.0.9
RapidFuzz==3.12.1

# --- Utilities & Logging ---
loguru==0.7.3
tenacity==9.0.0
tqdm==4.67.1
python-multipart==0.0.20

# --- Optional (Dev / Notebook / Debug) ---
ipykernel==6.29.5
nest-asyncio==1.6.0

# --- PDF, HTML, DOC Conversion ---
weasyprint==64.0
markdown==3.7
zopfli==0.2.3.post1

# For autonomous research agent
gpt-researcher==0.11.0
duckduckgo_search==7.3.2
google_search_results==2.4.2

